refactor(frontend): modularize map component for imporved scalibility
